refactor(): 调整代码组织，优化逻辑

+++ database/mysql/builder_delete.go
@@ -36,7 +36,13 @@ func (s *Builder) BuildDeleteRelation(vField model.Field, rModel model.Model) (d
 		return
 	}
 
-	relationTableName := s.common.GetRelationTableName(vField, rModel)
+	relationTableName, relationErr := s.common.GetRelationTableName(vField, rModel)
+	if relationErr != nil {
+		err = relationErr
+		log.Errorf("BuildDeleteRelation %s failed, s.common.GetRelationTableName error:%s", vField.GetName(), err.Error())
+		return
+	}
+
 	delRight = fmt.Sprintf("DELETE FROM `%s` WHERE `%s` IN (SELECT `right` FROM `%s` WHERE `left`=%v)",
 		s.common.GetModelTableName(rModel),
 		rModel.GetPrimaryField().GetName(),
